+++ pkg/utils/instance.go
@@ -2,6 +2,6 @@ package utils
 
 import "fmt"
 
-func GenerateInstanceName(clusterID string, nodeClaimName string) string {
-	return fmt.Sprintf("%s-%s", clusterID, nodeClaimName)
+func GenerateInstanceName(prefix string, nodeClaimName string) string {
+	return fmt.Sprintf("%s-%s", prefix, nodeClaimName)
 }